### 2019-07-25 #329 @RaoulWolf
chemspiderapi
Submitting Author Name: Raoul Wolf
Submitting Author Github Handle: @RaoulWolf
Repository:  https://github.com/RaoulWolf/chemspiderapi
Version submitted: 0.0.2  2021-03-15
Reviewers: @rajarshi, @yufree, @data-datum
Version accepted: TBD

---

```
Package: chemspiderapi
Title: R Wrapper for ChemSpider's API Services
Version: 0.0.2
Authors@R: person("Raoul", "Wolf", email = "raoul.wolf@niva.no", role = c("aut", "cre"))
Description: ChemSpider has announced a fundamental change to the syntax of their API services in late 2018.
    This package provides convenience wrappers for the new API functionalities of ChemSpider, as well as complementary functions.
    curl,
    jsonlite
    covr,
    keyring,
    magick,
    memoise,
    ratelimitr,
```

The `chemspiderapi` package is an easy-to-use R interface to use all new ChemSpider API functionalities, as introduced in ChemSpiders complete redesign of its API structure late 2018.

Researchers and citizen scientists who work on anything chemistry-related and need to routinely query against any of ChemSpider's API services.

The rOpenSci-maintained package webchem aims at offering (currently outdated / not working) functionality for accessing ChemSpider's API services. However, the redesign of ChemSpider's API structures in late 2018 broke all available functionalities.

Pre-submission enquiry: https://github.com/ropensci/software-review/issues/294
@melvidoni
