### 2019-07-25 #329 @RaoulWolf
Hi @annakrystalli, yes I submitted my fork on purpose. I couldn't get an OK to incorporate AppVeyor, CodeCov and Travis CI on the other repository, so I forked it into my own account and added the functionality.
