### 2019-07-25 #329 @annakrystalli
OK, that's fine. I was just wondering whether further development would be done in the upstream fork or yours?

Also, we require test coverage of at least 75% before review. Your codecov badge indicates only 10% test coverage. Is there a reason for that?
