### 2019-07-29 #329 @RaoulWolf
Further development would be at the upstream fork, yes. It looks like I might be able to get Circle CI and CodeCov up and running on the upstream fork. Would that be preferable then?

Coverage is an issue indeed, but comes with the nature of the package. I've increased the coverage considerably (>30%), but I'm not sure how to increase it further. The main issue is that I do not want the tests to run the actual API queries because there's a rate limit on the queries. Any recommendations are more than welcome :)
