### 2019-07-31 #329 @annakrystalli
Hey @RaoulWolf.

Thanks for the updates and your efforts with improving test coverage!

 Regarding testing the API, have you had a look at package [httptest](https://cran.r-project.org/web/packages/httptest/index.html)?  This should help with what you're trying to achieve.
