### 2019-08-05 #329 @sckott
@RaoulWolf  Another option for caching tests (not doing real requests) is the vcr package - but only will work if you use crul or httr instead of curl
