### 2019-08-09 #329 @RaoulWolf
Thanks @sckott for the heads up on `vcr`.

As you indicated, `vcr` (currently?) only works for `crul` or `httr`; also `httptest` does not work for `curl`. The reason I chose `curl` for the package was the relative flexibility over `httr` when assembling headers and data fields. In some instances I was not able to set up working API calls using `httr`.

I will take a further look into `crul` - of course I would prefer a solution where `httptest` or `vcr` would offer support for `curl` 😃

In the meantime, the upstream fork at NIVANorge (https://github.com/NIVANorge/chemspiderapi) has support for AppVeyor, CircleCi, Travis CI, and CodeCov. I assume it's fair game to change the repo address in the OP?
