### 2019-08-09 #329 @sckott
yeah `curl` is a great choice. jeroen is considering integration for webmockr (and therefore vcr) https://github.com/jeroen/curl/pull/174 but not sure if it will happen
