### 2019-08-16 #329 @RaoulWolf
Unfortunately I did not get `crul` to work for all possible functionalities; I have thus decided to stick with `curl`.

Otherwise I was able to bump up the test coverage to over 40% without using the actual API query functionalities 👍 I'm still eager to improve coverage, but I'm not sure how to at this moment...? Any help/recommendation is highly appreciated!

I will also update the OP to link to the upstream repository (https://github.com/NIVANorge/chemspiderapi).
