### 2019-08-16 #329 @noamross
There's also https://github.com/nealrichardson/httptest
