### 2020-01-24 #329 @maelle
@RaoulWolf, you can ask any question reg testing on https://discuss.ropensci.org :-)
