### 2020-04-22 #329 @stitam
Thank you @RaoulWolf for your work on chemspiderapi! I am one of the maintainers of webchem. I would like to note that since 2019 September webchem can also access the new ChemSpider web service (https://github.com/ropensci/webchem/issues/149), so I think there might be a package overlap issue (https://devguide.ropensci.org/policies.html#overlap). We would be very happy if this could be resolved, please contact us if you wish to dicuss it in more detail.
