### 2020-10-05 #329 @RaoulWolf
@maelle Sincerest apologies for the late reply - I haven't received any notifications regarding this submission until now! All answers since November 2019 are new to me... Oh my!

I am currently on vacation but I'm still very keen on pursuing `chemspiderapi` further. We have been using it routinely in our work, and last time I checked it provided more functionality than what was (is?) available over at `webchem`.

I haven't checked the progress with regard to possible API testing (as I haven't received notifications... again apologies!), but if you'd be willing to re-open this submission I'd be happy to do so.
