### 2020-10-05 #329 @stitam
Hi @RaoulWolf, I am a maintainer of `webchem`. Please note `webchem` is in heavy development, we have had three releases in the past 12 months including v1.0.0. The ChemSpider API has been fixed for about a year now and I think your package has significant overlap with `webchem`, only `webchem` provides access to other webservices as well. A paper using the fixed ChemSpider API in `webchem` has recently been published (https://www.jstatsoft.org/article/view/v093i13). Happy to discuss.
