### 2020-10-06 #329 @maelle
:wave: @RaoulWolf! Seeing that webchem wraps the ChemSpider API again I am a bit wary of [overlap](https://devguide.ropensci.org/policies.html#overlap). Is there anything webchem does not support?
