### 2020-10-06 #329 @RaoulWolf
@maelle thank you for re-opening the issue! Appreciated

@stitam very valid point. I'll get back to this in about a week from now. Thanks for the heads-up! (and for the record, I'm a `webchem` fan 😉)
