### 2020-10-19 #329 @RaoulWolf
Lengthy post ahead, so watch out!

I took some time to compare the solutions of `webchem` and compare them to what is offered in `chemspiderapi`. The first step to adress the potential big overlap is to compare the available functionality, starting with everything ChemSpider offers (https://developer.rsc.org/docs/compounds-v1-trial/1/overview), and comparing the availability between `chemspiderapi` and `webchem` *(I tried to be as thorough as possible, but please forgive me if I oversaw functionalities within `webchem`!)*:

**FILTERING**

| ChemSpider compound API                  | `chemspiderapi` wrapper               | `webchem` wrapper|
| :--------------------------------------- | :------------------------------------ | :-----------------------: |
| filter-element-post                      | `post_element()`                      |            ?            |
| filter-formula-batch-post                | `post_formula_batch()`                |            ?            |
| filter-formula-batch-queryId-results-get | `get_formula_batch_queryId_results()` |            ?            |
| filter-formula-batch-queryId-status-get  | `get_formula_batch_queryId_status()`  |           ?            |
| filter-formula-post                      | `post_formula()`                      |            `cs_formula_csid()`            |
| filter-inchi-post                        | `post_inchi()`                        |            `cs_inchi_csid()`            |
| filter-inchikey-post                     | `post_inchikey()`                     |            `cs_inchikey_csid()`            |
| filter-intrinsicproperty-post            | `post_intrinsicproperty()`            |            ?            |
| filter-mass-batch-post                   | `post_mass_batch()`                   |            ?            |
| filter-mass-batch-queryId-results-get    | `get_mass_batch_queryId_results()`    |            ?           |
| filter-mass-batch-queryId-status-get     | `get_mass_batch_queryId_status()`     |            ?            |
| filter-mass-post                         | `post_mass()`                         |            ?           |
| filter-name-post                         | `post_name()`                         |           `cs_name_csid()`            |
| filter-queryId-results-get               | `get_queryId_results()`               |            `cs_query_csid()`            |
| filter-queryId-results-sdf-get           | `get_queryId_results_sdf()`           |            ?           |
| filter-queryId-status-get                | `get_queryId_status()`                |            ?            |
| filter-smiles-post                       | `post_smiles()`                       |            `cs_smiles_csid()`           |

**LOOKUPS**

| ChemSpider compound API | `chemspiderapi` wrapper | `webchem` wrapper |
| :---------------------- | :---------------------- | :-----------------------: |
| lookups-datasources-get | `get_datasources()`     |            `cs_datasources()`            |

**RECORDS**

| ChemSpider compound API                 | `chemspiderapi` wrapper             | `webchem` wrapper |
| :-------------------------------------- | :---------------------------------- | :-----------------------: |
| records-batch-post                      | `post_batch()`                      |            ?             |
| records-recordId-details-get            | `get_recordId_details()`            |            `cs_compinfo()`/`cs_extcompinfo()`            |
| records-recordId-externalreferences-get | `get_recordId_externalreferences()` |            ?             |
| records-recordId-image-get              | `get_recordId_image()`              |            `cs_img()`             |
| records-recordId-mol-get                | `get_recordId_mol()`                |            ?             |

**TOOLS**

| ChemSpider compound API      | `chemspiderapi` wrapper    | `webchem` wrapper |
| :--------------------------- | :------------------------- | :-----------------------: |
| tools-convert-post           | `post_convert()`           |           `cs_convert()`/`cs_convert_multiple()`             |
| tools-validate-inchikey-post | `post_validate_inchikey()` |            ?             |

While there (quite unsurprisingly) is overlap, `chemspiderapi` offers direct access to *all* API functionalities of ChemSpider.

Another major difference between the solution offered by `chemspiderapi` and `webchem` is the workflow. It's really two different philosophies in my view, so there's no right or wrong. `chemspiderapi` is very explicit (and maybe lengthy) in every step, but "forces" users to familiarize themselves with the desired API use workflow of ChemSpider; additionally, `chemspiderapi` offers several vignettes to guide users in ways how to store their API token(s), rate-limiting quires, memoising queries, and saving .mol, .sdf, or .png files. `webchem` has an elegant "under the hood" approach to the ChemSpider's API functionalities it offers.

Maybe less interesting, but also worth noting, is the unsurprising difference in dependencies. `chemspiderapi` has two dependencies, and `webchem` has 13 dependencies. `chemspiderapi` was designed to have as little dependencies as possible, and both dependencies (`curl` and `jsonlite`) are well-established and introduce no other dependencies.

While for most users the functionalities within `webchem` are likely enough, I still think `chemspiderapi` provides a more "total" ChemSpider API experience, for anyone wishing to go down this rabbit hole.

Let's discuss!
