### 2020-10-19 #329 @maelle
Thanks a lot for your detailed answer! Could you please add a column or sentences explaining to humans what e.g. _records-batch-post_ does i.e. the type of functionalities your package adds? I have the impression your package allows for posting information, what are the use cases for that? And what about the functionalities it adds that have nothing to do with posting?

Thanks again!
