### 2020-10-19 #329 @RaoulWolf
Right, apologies for not providing enough context!

The different `post_*()` and `get_*()` functionalities are named so from the HTTP methods ([POST](https://en.wikipedia.org/wiki/POST_(HTTP)) and [GET](https://en.wikipedia.org/wiki/GET_(HTTP)), respectively).

In simple terms, all `post_*()` functions "upload" information into ChemSpider's API services (and usually get a query ID in return), and all `get_*()` functionalities "download" information from ChemSpider's API services.

In the example case you mentioned (*records-batch-post*), a list of up to 100 ChemSpider record IDs is POST-ed to ChemSpider's API services, and a query ID is returned.

I hope this answers your question!
