### 2020-10-20 #329 @maelle
Thanks, and in what kinds of workflows do you need to post information?

@stitam could you please confirm the question marks in the comparison table, i.e. that webchem does not provide that functionality?
