### 2020-10-20 #329 @RaoulWolf
The principle workflows for the filtering functionalities all follow the same pattern:

* POST query,
* GET status, and
* GET results.

This is also mentioned in the [README](https://github.com/NIVANorge/chemspiderapi/blob/master/README.md) of `chemspiderapi`. The lookups, records and tools functionalities work directly, i.e., without the (manual) three step process.
