### 2020-10-20 #329 @stitam
Hi All! @RaoulWolf thanks for your detailed answer to @maelle's questions.

@RaoulWolf is right in saying that the two packages follow different philosophies and support different types or workflows. My understanding is that `chemspiderapi` aims to develop an R function for each ChemSpider API. On the other end, `webchem` focuses on the user experience, and aims to distance the user from the API, so instead of being "process" focused it is more "outcome" focused. I agree with @RaoulWolf in that there is no good or bad, these are two different approaches.

It follows from the difference in philosophies that `webchem` doesn't implement all APIs. The beginning of a `webchem` workflow is that a user is looking for a "specific" compound or set of compounds. Therefore we skipped a few APIs like `filter-mass-post` where the user would specify a molecular weight range and the API (actually 2-3 APIs called after eachother) would return a list of CSIDs for compounds with molecular weights falling in that range. At this point we didn't find a good enough use case for that functionality and our user haven't asked for it either.

We also didn't implement functions that process batch requests. We do see the benefit of batch requests, but not all requests have a batch alternative at the moment, so it was difficult to bind them all into a simple and easy to use user facing function. We did implement the non-batch alternatives for these queries however. ChemSpider APIs seem to be under development so once batch is available for all the queries we need (it's actualy `filter-name-batch-post` we are missing) then we'll add the batch options as well to make our queries more efficient.

Finally a few notes on the table above, in terms of user experience, in `webchem` all ChemSpider APIs that ultimately return CSIDs are actually pooled into `get_csid()`, that is the exported function. Also `get_queryId_status` is called within the functions to query the status before requesting the response itself. Other than these, yes @maelle, I can confirm the question marks, those APIs are currently not implemented in `webchem` so there is no overlap there at the moment.
