### 2020-10-25 #329 @jooolia
Editor comments:

Hi @RaoulWolf

Thank you for all of the comments and answers to our questions that you have provided. We will move ahead with the onboarding process for this package. I will be the editor handling the process going forward.

- We have been happy to see the discussion with the maintainers of {webchem} and think that it would be useful to mention webchem in your documentation. It is also a hope that these packages could work well together in the future (i.e. perhaps chemspiderapi could become a dependency for webchem or some other scenario).
 - We do [require 75 % code coverage for a package to go to review]((https://devguide.ropensci.org/building.html#testing) (and currently the code coverage is at 41% for this package), but I understand there is some difficulty with testing functions using an API. Some of the other editors have suggested [{presser}](https://r-lib.github.io/presser/) as a way to help create some of these tests.

 Here are the results of some of the issues flagged with some of our initial editor checks:

```
── GP chemspiderapi

    R/FILTERING-get_formula_batch_queryId_results.R:29:NA
    R/FILTERING-get_formula_batch_queryId_results.R:31:NA
    R/FILTERING-get_formula_batch_queryId_results.R:33:NA
    R/FILTERING-get_formula_batch_queryId_results.R:35:NA
    R/FILTERING-get_formula_batch_queryId_results.R:37:NA
```

- [ ] increase code coverage of unit tests
```

  ✖ add a "BugReports" field to DESCRIPTION, and point it to a bug tracker. Many online code hosting services provide
    bug trackers for free, https://github.com, https://gitlab.com, etc.

```

- [ ]  add a "BugReports" field to description

```
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80 characters

    R/CHECKING-check_apikey.R:8:1
    R/CHECKING-check_apikey.R:12:1
    R/CHECKING-check_apikey.R:16:1
    R/CHECKING-check_complexity.R:8:1
    R/CHECKING-check_elements.R:12:1

```
- [ ] check code line length (where possible)

I will look for package reviewers once these issues have been addressed. Please let me know if you have any questions or if I can clarify anything.

Thanks! Julia
