### 2020-10-26 #329 @RaoulWolf
Hi @jooolia, thanks for the update! I'll try to fix the easier issues (bug reports and code line length) this week, and tackle code coverage next week. I haven't checked `presser` yet, thanks for the heads up! I'll let you know once the issues are addressed.

Thanks for the effort! Raoul
