### 2020-10-27 #329 @RaoulWolf
The `BugReports` field was added to the description, the `README` now mentions `{webchem}`, and I tried to minimize the line lengths as much as possible.

All changes are live at https://github.com/NIVANorge/chemspiderapi

Code coverage hasn't been addressed yet.
