### 2020-11-09 #329 @RaoulWolf
A small update regarding test coverage.

I tried implementing API tests with both {httptest} and {vcr}, but with no luck. {vcr} does not seem to support {curl} just yet (the [introduction](https://cran.r-project.org/web/packages/vcr/vignettes/vcr.html) hints at future support), and {httptest} simply runs the actual queries without mocking.

I am now taking a look at {presser} and will let you know how far I get.
