### 2020-11-09 #329 @maelle
{presser} supports curl indeed and that's the only HTTP testing package that does. :-)
