### 2020-11-17 #329 @RaoulWolf
Small update: I finally managed to wrap my head around {presser} and the first tests are running (see [here](https://github.com/NIVANorge/chemspiderapi/blob/master/tests/testthat/test-get_datasources.R), line 77). I'll now try to extend the testing functionality to the other functions. Baby steps 😃

Meanwhile, I've also added GitHub Actions as CI (trying to substitute CircleCI and Travis CI). Unfortunately the {curl} requirement keeps breaking the installation on Linux machines. My initial approach was adding the following chunk to `R-CMD-check.yaml`:

````
     - name: Install libcurl
        if: runner.os == 'Linux'
        run: |
          sudo apt-get install -y libcurl4-openssl-dev
````

The test then fails when (trying to) installing {chemspiderapi}:
````
Error in dyn.load(file, DLLpath = DLLpath, ...) :
  unable to load shared object '/home/runner/work/_temp/Library/curl/libs/curl.so':
  /usr/lib/x86_64-linux-gnu/libcurl.so.4: version `CURL_OPENSSL_3' not found (required by /home/runner/work/_temp/Library/curl/libs/curl.so)
````

I also tried with `libcurl4-gnutls-dev`, but no luck either. Any tips warmly welcome!
