### 2020-11-17 #329 @maelle
Untested tip, I'd look at what other curl dependencies have in their workflows e.g. https://github.com/r-lib/httr/blob/cb4e20c9e0b38c0c020a8756db8db7a882288eaf/.github/workflows/R-CMD-check.yaml#L60
