### 2020-11-17 #329 @RaoulWolf
Turns out the "trick" is to not use Ubuntu 20.xx at all, but good ol' 16.04. Now all tests pass 👍
