### 2020-11-25 #329 @RaoulWolf
Hi @jooolia, It first looked like {presser} would help and I can definitely setup mock-APIs with it (hurray!). But I still struggle to increase coverage. I suppose it's because the API request itself happens within {chemspiderapi} functions, but I'm not sure. An example can be seen [here](https://github.com/NIVANorge/chemspiderapi/blob/master/tests/testthat/test-get_datasources.R), line 77. The tests pass, but the coverage is not increased, which is slightly frustrating. On the positive side I have now over 500 tests for the package that pass 😅

For the CI I was wondering if GitHub actions would be enough? I guess there's no need for Travis, AppVeyor or CircleCI?
