### 2020-11-25 #329 @maelle
Commenting on this again, hope it's fine. The presser package now has a function with which you don't need to do setup/teardown
* https://r-lib.github.io/presser/dev/reference/local_app_process.html
* https://r-lib.github.io/presser/dev/articles/how-to.html#how-do-i-start-the-app-when-writing-the-tests-
