### 2020-11-25 #329 @jooolia
Hi @RaoulWolf,
Yes the CI using GitHub actions should be sufficient as you have it implemented.

I am also trying to wrap my head around webfakes  (previously presser now at https://r-lib.github.io/webfakes/ ) so I cannot help a lot with that aspect (but I will try as I am curious about how to do this sort of testing), however one thing that I see when I run `covr::report()` is that there are many arguments in the functions that are not tested in the tests so these lines are never run and thus not covered.  And this is also true with the API testing, since none of the functions from the package are called in the mock API the coverage does not increase.

I will try to look a bit more at webfakes over the next few days and see if I can help.

@maelle your inputs and wisdom are always welcome. :)
