### 2020-11-25 #329 @jooolia
(Just putting this here for reference, there is a nice book by @maelle about http testing: https://books.ropensci.org/http-testing/packages-for-http-testing.html)
