### 2020-11-26 #329 @maelle
Thanks @jooolia :blue_heart: Demo of webfakes at https://github.com/ropensci-books/http-testing/pull/47 with demos of other packages.

So it seems the low coverage is not a webfakes problem, correct? I'm happy to help if needed.
