### 2020-11-27 #329 @RaoulWolf
Hi @jooolia and @maelle, very cool to see {presser} become {webfakes}! I'll keep an eye out on the book 😃

I'll try a few more variants to increase coverage with {webfakes} over the weekend; maybe there's a way I haven't thought about...
