### 2020-12-08 #329 @RaoulWolf
I tried now several other ways to code tests so it can run with "mock" APIs using {webfakes}, but I simply cannot find a way to emulate a (hardcoded URL) API from outside the function which defines the API itself.

All other tests work well, but this seems very complicated to put in place at this point. I looked into {webchem}'s [approach](https://github.com/ropensci/webchem/blob/master/tests/testthat/test-chemspider.R) to this, and it seems they - apart from the very reasonable `skip_on_cran()` and `skip_on_ci()` controls - actually run real queries.

Given my current subscription with ChemSpider I'd very much like to avoid running alot of queries every time the package is updated/revised.

How do we proceed now? I'm very much open to give {webfakes} another try with some help, but otherwise I seem to be stuck in a dead-end when it comes to increasing test coverage.
