### 2020-12-08 #329 @maelle
Where is that URL hard-coded? Could you use an internal function to define the API URL instead as in https://books.ropensci.org/http-testing/webfakes.html#actual-testing-2?

Reg real requests we've added this chapter to the book https://books.ropensci.org/http-testing/real-requests-chapter.html
