### 2020-12-08 #329 @RaoulWolf
The respective URLs are hard-coded in the respective functions, for example [here](https://github.com/NIVANorge/chemspiderapi/blob/master/R/FILTERING-post_inchikey.R):
````
url <- "https://api.rsc.org/compounds/v1/filter/inchikey"
````
This URL (and all others) are then used in `curl::curl_fetch_memory()`. Most of the API functionalities come with dedicated URLs, so this way of hard-coding allows to easier keep track of the URLs, in case there are any future changes.

Additionally, it is sometimes necessary to modify the hard-coded URL depending on user-defined input, such as [here](https://github.com/NIVANorge/chemspiderapi/blob/master/R/FILTERING-get_queryId_results.R), where the minimum requirement is to provide a `queryId`, and `start` and `count` are optional:
````
if (is.null(start) && is.null(count)) {
    url <- paste0("https://api.rsc.org/compounds/v1/filter/", queryId, "/results")
  }

  if (!is.null(start) && is.null(count)) {
    url <- paste0("https://api.rsc.org/compounds/v1/filter/", queryId, "/results?start=", start)
  }

  if (is.null(start) && !is.null(count)) {
    url <- paste0("https://api.rsc.org/compounds/v1/filter/", queryId, "/results?count=", count)
  }

  if (!is.null(start) && !is.null(count)) {
    url <- paste0("https://api.rsc.org/compounds/v1/filter/", queryId, "/results?start=", start, "&count=", count)
  }
````
I'll take a look into the "Making real requests" chapter of the book asap! This makes me wonder if there's a sensible way to run "real" API tests without exceeding rate-limitations; possibly a brute-force use of `Sys.sleep()` could help? Then again, if the tests are excluded from running on CI systems (`skip_on_ci()`), the testing coverage will not increase either...
