### 2020-12-08 #329 @maelle
Instead of having

```r
url <- "https://api.rsc.org/compounds/v1/filter/inchikey"
```

you could have

```r
url <- get_compounds_inchikey_url()
```

where `get_compounds_inchikey_url()` is

```r
get_compounds_inchikey_url <- function() {
  api_url <- Sys.getenv("COMPOUNDS_INCHIKEY_URL")

  if (nzchar(api_url)) {
    return(api_url)
  }

  "https://api.rsc.org/compounds/v1/filter/inchikey"
}
```

This way you can override the URL by setting the environment variable to the URL of a webfakes app process.
