### 2020-12-08 #329 @maelle
Same for your other snippet, you'd replace "https://api.rsc.org/compounds/v1/filter/" with `get_compounds_filter_url()`. Does it make sense?
