### 2020-12-08 #329 @maelle
> This makes me wonder if there's a sensible way to run "real" API tests without exceeding rate-limitations; possibly a brute-force use of Sys.sleep() could help? Then again, if the tests are excluded from running on CI systems (skip_on_ci()), the testing coverage will not increase either...

You'd make sure not to run them too often. Another thing that might help, or not, is asking the web API providers if you can get some sort of special developer account (I mention this in the book, for some APIs it can help as package maintainers are helping spread their product/API, after all).

True about test coverage, hopefully by making it possible to override the API URLs you can use more webfakes apps.
