### 2020-12-08 #329 @maelle
Also, if you don't mind my asking: why do you use curl by the way? (rather than httr or crul)
