### 2020-12-09 #329 @RaoulWolf
Lot's to unpack here, thank you for the detailed recommendations @maelle! Let's start from the end:

> Also, if you don't mind my asking: why do you use curl by the way? (rather than httr or crul)

when I first started working on the package, {curl} was the only package that would allow to run all different flavors of API queries that were necessary. For the record, in the beginning I tried implementing the functionalities with {httr} and {crul}, but not all API queries were possible. Apart from that {curl} comes with no dependencies, which I personally prefer (that's also the reason why I use {jsonlite} in the package).

> Another thing that might help, or not, is asking the web API providers if you can get some sort of special developer account

...let's just say we should explore other options first ;)

> This way you can override the URL by setting the environment variable to the URL of a webfakes app process.

I like the idea of using environmental variables, {webchem} uses it as well if I'm not mistaken. My worry is that the readability (and, after all, maintenance) of the code becomes more difficult with these constructs. To me this seems like a trade-off situation between code simpleness/tidiness on one hand and increasing test coverage on the other hand. But I think these thoughts are a bit bigger than my humble package :) I'll try to see if this approach works!
