### 2020-12-11 #329 @maelle
A more compact version would be

```r
baseurl <- function() {
  Sys.getenv("COMPOUNDS_INCHIKEY_URL", "https://api.rsc.org/compounds/v1/filter/inchikey")
}
```

Reg https://github.com/NIVANorge/chemspiderapi/blob/master/R/FILTERING-get_queryId_results.R?rgh-link-date=2020-12-08T14%3A32%3A34Z you might want to create a helper function that returns the URL depending on the variables (to have less if/else)?
