### 2021-01-06 #329 @RaoulWolf
I now implemented {webfakes} for most API calls and the test coverage is around 97%. CI is now fully switched to Github Actions. Everything is live at the repository: https://github.com/NIVANorge/chemspiderapi

Anything else needed?
