### 2021-03-04 #329 @yufree
need extra work to fix some errors.
- [ ] **Examples** (that run successfully locally) for all exported functions
    need to fix some issues.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Overall it's a useful package for researchers to retrieve information from ChemSpider API. I tested this package as a chemical researcher and software developer. As a researcher, I might need the example of every function could return results without errors. As a software developer, I would check if the package have the proper style and readable code for other R users. At current stage, this package still need some improvements.

#### Information about Chemspider API

It would be helpful to mention that free account will have up to 1000 API requests per month in the readme.

#### `goodpractice` and `spelling` check

For `goodpractice` check, lots of code lines were longer than 80 characters. I checked those lines and found some long code lines might need to format in a shorter way for the reader. Developer can consider `Reformat Code` function in RStudio to improve the codes. Another issues were about the test coverage, which looks fine for me.

For `spelling` check, most of the results were terms in Cheminformatics, which should be fine. However, in Rate-Limiting_API_Queries.Rmd:21, 'recomendations' should be 'recommendations'.

#### `inst` folder

I found the files in the `inst` folder and I wonder if the developer could mention this file somewhere for the user as a short introduction of this package. Or the developer could consider to move the contents to `<package-name>.R` and use `roxygen2` package to generate the Rd file for the whole package.

#### Function example

I think something might changed during the past year and some examples of function will not work. Because most of examples were protected by do-not-run with API token, regular test will not find those issues. I hope the developer could test those examples in the revision.

- The major issue in the example is that status is a list object and it's always needed to get the third element. In this case, `status=status` should be changed to `unlist(status)[3]` to get the 'Complete' status. Almost every function's example need to change this part to make sure the example could be run.

- Examples of `get_formula_batch_queryId_status`, `get_mass_batch_queryId_results`, `get_queryId_results_sdf` return error

- Example of `get_recordId_externalreferences` should return results instead of Null list for the user

#### vignette: Rate-Limiting API Queries

Line 91. "All queries need to report the status `Complete` before proceeding." and unlist(y)[1] only return one element as 1. The third element is the 'Complete' and that will give the chemspider ID. I wonder if this issue is from the change of API? Meanwhile, the returned object should be a vector of chemspider ID. This line could be changed to

```r
demo_chemicals$id = mapply(FUN = function(x, y) sleepy_get_queryId_results(queryId = x, status = unlist(y)[3], apikey = apikey, coerce = T)$results, x = demo_chemicals$queryId, y = demo_chemicals$status)
```
Line 151. Similar issue as line 91

Line 189. Here the code should be

```r
mutate(queryId = map_chr(InChIKey, ~ slowly_post_inchikey(inchikey = .x, apikey = apikey)$queryId))
```

Otherwise user will get the error "Error: Problem with `mutate()` input `queryId`.
x Can't coerce element 1 from a list to a character
ℹ Input `queryId` is `map_chr(InChIKey, ~slowly_post_inchikey(inchikey = .x, apikey = apikey))`."

Line 200. The code should be

```r
mutate(status = map(queryId, ~ unlist(slowly_get_queryId_status(queryId = .x, apikey = apikey)[3]))) %>%
```

Line 212. The code should be

```
mutate(id = map2_int(queryId, status, ~ slowly_get_queryId_results(queryId = .x, status = .y, apikey = apikey)$results))
```

#### vignette: Remembering API Queries

This is an empty vignette. Maybe remove this till it's finished.

#### vignette: Saving MOL and SDF Files

Line 32-34. Maybe remove this chunk.

Line 44. Should be `writeLines(caffeine_mol$sdf, con = con)`

Line 54. Should be `queryId <- post_mass(mass = 194, range = 0.002, apikey = apikey)$queryId`

Line 66. Should be `caffeine_sdf <- get_queryId_results_sdf(queryId = queryId, status = status$status, apikey = apikey)`

#### vignette: Saving PNG Images

Actually `png` package could do this work. Maybe we don't need `magick` package unless we need extra features.

```r
caffeine_image_raw <- get_recordId_image(recordId = 2424L, apikey = apikey,decode = T)
png::writePNG(png::readPNG(caffeine_image_raw$image), "mypng.png")
```

In general, it's nice to see a package for ChemSpider API. As discussed above, `webchem` might have overlap functions while `chemspiderapi` give access to all of the APIs provided by ChemSpider. I hope those mentioned issues could be fixed.

Hope you will find the comments useful.

Thanks for the review invitation.

Miao
