### 2021-03-22 #329 @rajarshi
- [X] **Examples** (that run successfully locally) for all exported functions

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

The package provide a consistent way to access Chemspider API endpoints. As already noted, there is overlap with `webchem`, and given the more expansive nature of this package, it may be worth considering merging this into `webchem` so that users can go to one place to access ChemSpider from R

It would be useful to note in the README  a link to the RSC dev pages where a user can get an API key

I'm not a fan of the combination of camelCase and underscore's in function names. For example `get_recordId_mol` could be `get_record_id_mol` or `getRecordIdMol` (with the former being preferred).

The [Remembering API Queries](https://github.com/NIVANorge/chemspiderapi/blob/master/vignettes/Remembering_API_Queries.Rmd) vignette seems to be empty

The rate limiting vignette is a great example of responsibly interacting with web APIs

Having to specify the api key with every call seems inefficient. One way to avoid this would be to have an `env` that is not exported, so that a user could set the key at the beginning of a session, which would add the key to the `env` and then functions would internally get the value of the key from the env. It would be good practice to have the `apikey` argument still present but with a default of `NULL`, so that if the user wanted to temporarily use a different key, they could.

Overall a well written package, that provides a comprehensive interface to the ChemSpider API
