### 2021-03-29 #329 @data-datum
Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide

- [X] Briefly describe any working relationship you have (had) with the package authors.

As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor before starting your review).

- [X] A statement of need clearly stating problems the software is designed to solve and its target audience in README

- [X] Installation instructions: for the development version of package and any non-standard dependencies in README

- [ ] Vignette(s) demonstrating major functionality that runs successfully locally need extra work to fix some errors.

- [X] Function Documentation: for all exported functions

- [ ] Examples (that run successfully locally) for all exported functionsneed to fix some issues.

- [ ] Community guidelines including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).

> For packages co-submitting to JOSS
>
> The package has an obvious research application according to JOSS's definition
>
> The package contains a paper.md matching JOSS's requirements with:
>
> A short summary describing the high-level functionality of the software
> Authors: A list of authors with their affiliations
> A statement of need clearly stating problems the software is designed to solve and its target audience.
> References: with DOIs for all those that have one (e.g. papers, datasets, software).
>

- [X] Installation: Installation succeeds as documented.

- [X] Functionality: Any functional claims of the software been confirmed.

- [ ] Performance: Any performance claims of the software been confirmed.

- [ ] Automated tests: Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

- [ X] Packaging guidelines: The package conforms to the rOpenSci packaging guidelines

* The “Remembering API Queries” vignette is empty.
* The examples are not clear. It has a legend “Not Run”.
* It would be great to add a CONTRIBUTING section.
