### 2021-06-18 #329 @RaoulWolf
Hi again @jooolia , and sorry for the delay. I've started working through the comments of the reviewers and have addressed most of the issues already - I'll update this thread once the revision is complete.

Meanwhile, the main repository for the package will be [RaoulWolf/chemspiderapi](https://github.com/RaoulWolf/chemspiderapi) after all (and not the previous repository); from August on I will not have access to the previous repository anymore and it will likely become private. Should the "new" repository ([RaoulWolf/chemspiderapi](https://github.com/RaoulWolf/chemspiderapi)) be a fork of the old one (like it is now) or rather a fresh repository?
