+++ tests/testthat/test-get_mass_batch_query_id_results.R
@@ -1,0 +1,155 @@
+library(chemspiderapi)
+
+context("get_mass_batch_query_id_results")
+
+test_that("get_mass_batch_query_id_results() fails if no query_id is provided.", {
+  expect_error(
+    get_mass_batch_query_id_results()
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if a NULL query_id is provided.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = NULL)
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if a query_id is not a character vector.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = 123)
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if multiple query_ids are provided.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = c("0c98889f-5e8b-4974-aabb-31ab28c54261", "0c98889f-5e8b-4974-aabb-31ab28c54261"))
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if a query_id is too long.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8b-4974-aabb-31ab28c54261N")
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if a query_id is not hyphen divided into five parts.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8b-4974-aabb31ab28c54261N")
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if the first part a query_id is of the wrong length.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889fN-5e8-4974-aabb-31ab28c54261")
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if the second part a query_id is of the wrong length.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8bN-4974-aab-31ab28c54261")
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if the third part a query_id is of the wrong length.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8b-4974N-aab-31ab28c54261")
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if the fourth part a query_id is of the wrong length.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8b-4974-aabbN-31ab28c5426")
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if no status is provided.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8b-4974-aabb-31ab28c54261")
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if a NULL status is provided.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8b-4974-aabb-31ab28c54261", status = NULL)
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if a multiple status are provided.", {
+  expect_error(
+    get_mass_batch_query_id_results(query_id = "0c98889f-5e8b-4974-aabb-31ab28c54261", status = c("Complete", "Complete"))
+  )
+})
+
+test_that("get_mass_batch_query_id_results() fails if a non-character status is provided.", {
+  expect_error(
[…]
