+++ README.md
@@ -2,6 +2,7 @@ chemspiderapi
 ================
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
 <!-- badges: start -->
 
 [![R build
@@ -10,23 +11,23 @@ status](https://github.com/NIVANorge/chemspiderapi/workflows/R-CMD-check/badge.s
 coverage](https://codecov.io/gh/NIVANorge/chemspiderapi/branch/master/graph/badge.svg)](https://codecov.io/gh/NIVANorge/chemspiderapi?branch=master)
 <!-- badges: end -->
 
-*As of 1 June 2021 this is the new official repository of
+*As of 18 June 2021 this is the new official repository of
 {chemspiderapi}. The old repository is orphaned.*
 
 > R functionalities for ChemSpider’s API services
 
 ChemSpider has introduced a new API syntax in late 2018, and [the old
-ChemSpider API syntax will be shut down at the end of November
-2018](http://link.rsc.org/rsps/m/xSq8Cm8ovjN8-Elm0eYB3Sey61zutqNIUUaMcyc14sQ).
+ChemSpider API syntax will be shut down at the end of
+November 2018](http://link.rsc.org/rsps/m/xSq8Cm8ovjN8-Elm0eYB3Sey61zutqNIUUaMcyc14sQ).
 `{chemspiderapi}` provides R wrappers around the new API services from
 ChemSpider.
 
 The aim of this package is to:
 
-1.  Translate the new ChemSpider API services into R-friendly functions.
-2.  Include thorough quality checking *before* the query is send, to
+1)  Translate the new ChemSpider API services into R-friendly functions.
+2)  Include thorough quality checking *before* the query is send, to
     avoid using up the query quota on, e.g., spelling errors.
-3.  Implement the R functionality in a way that is suitable for both
+3)  Implement the R functionality in a way that is suitable for both
     `{base}`- and `{tidyverse}`-style programming.
 
 `{chemspiderapi}` relies on API keys to access ChemSpider’s API
@@ -99,13 +100,13 @@ If `{curl}` or `{jsonlite}` are missing, (almost) all functions of
 
 ## Coverage
 
-As of 2021-06-07, the following functionalities are implemented (100%
-functionality with 100% annotation):
+As of 2021-06-18, the following functionalities are implemented and
+fully annotated (100%):
 
 **FILTERING**
 
 | ChemSpider Compound API                  | `{chemspiderapi}` Wrapper              |
-|:-----------------------------------------|:---------------------------------------|
+| :--------------------------------------- | :------------------------------------- |
 | filter-element-post                      | `post_element()`                       |
 | filter-formula-batch-post                | `post_formula_batch()`                 |
 | filter-formula-batch-queryId-results-get | `get_formula_batch_query_id_results()` |
@@ -127,13 +128,13 @@ functionality with 100% annotation):
 **LOOKUPS**
 
 | ChemSpider Compound API | `{chemspiderapi}` Wrapper |
-|:------------------------|:--------------------------|
+| :---------------------- | :------------------------ |
 | lookups-datasources-get | `get_data_sources()`      |
 
 **RECORDS**
 
 | ChemSpider Compound API                 | `{chemspiderapi}` Wrapper             |
-|:----------------------------------------|:--------------------------------------|
+| :-------------------------------------- | :------------------------------------ |
 | records-batch-post                      | `post_batch()`                        |
 | records-recordId-details-get            | `get_record_id_details()`             |
 | records-recordId-externalreferences-get | `get_record_id_external_references()` |
@@ -143,7 +144,7 @@ functionality with 100% annotation):
 **TOOLS**
 
 | ChemSpider Compound API      | `chemspiderapi` Wrapper    |
-|:-----------------------------|:---------------------------|
[…]
