### 2021-06-21 302d076
Removal of coerce and simplify functionality, rewrite of tests and simplification of code
