+++ .github/workflows/R-CMD-check.yaml
@@ -62,7 +62,7 @@ jobs:
       - name: Install Linux system dependencies
         if: runner.os == 'Linux'
         run: |
-          sudo apt-get install libcurl4-openssl-dev
+          sudo apt-get install -y libcurl3-openssl-dev libcurl4-openssl-dev
         
       - name: Install macOS system dependencies  
         if: runner.os == 'macOS'
