+++ .github/workflows/R-CMD-check.yaml
@@ -28,8 +28,6 @@ jobs:
           - {os: macOS-latest, r: 'release'}
           - {os: ubuntu-16.04, r: '3.5', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
           - {os: ubuntu-16.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
-          - {os: ubuntu-20.04, r: '3.5', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
-          - {os: ubuntu-20.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
@@ -62,7 +60,7 @@ jobs:
       - name: Install Linux system dependencies
         if: runner.os == 'Linux'
         run: |
-          sudo apt-get install -y libcurl3-openssl-dev libcurl4-openssl-dev
+          sudo apt-get install libcurl4-openssl-dev
         
       - name: Install macOS system dependencies  
         if: runner.os == 'macOS'
