### 2021-06-23 #329 @jooolia
Hi @RaoulWolf ,
Thanks for keeping us updated on your progress.

I do not see any reason that the fork repository would be problematic. I will update the information at the top of this issue with this new repo. After the review process is completed we will transfer it to the rOpenSci organization. @maelle , do you think there would be any problems if we need to transfer a repository that is a fork (instead of the original)?
