### 2021-06-29 03d60da
Updating the vignettes and checking code length
