### 2021-08-31 #329 @jooolia
Hi @RaoulWolf,
Do you have any updates on your revisions? If you are not able to work on it right now we could  put the "holding" label here. https://devguide.ropensci.org/policies.html?q=hold#policiesreviewprocess

Thanks! Julia
