### 2021-11-01 #329 @jooolia
Email sent to raoul.wolf@hotmail.de and raoul.wolf@niva.no asking for an update. Otherwise I will close this issue if we do not receive a response in a week( by 8 November 2021).
