### 2021-11-01 #329 @jooolia
Response to email received today:

> Hi Julia,
>
> thanks for reaching out! I've been quite busy the past few months (new job...) but finished most of the requested changes already, except a revised vignette. I hope to be able to dedicate some more time to it before Christmas, so a "holding" label would be fantastic.
>
> Once again, thank you for reaching out and have a great day!

Thus I will add the holding label until 2022 and then we can revisit.
