### 2022-06-10 #329 @jooolia
Email sent to raoul.wolf@hotmail.de

> Hi Raoul,
I am reaching out again regarding chemspiderapi. Do you have any updates? I think I will close the issue and if you have time to complete it you could reopen the issue and ping me on it.

>Thanks, Julia

With that I will close this issue. Cheers, Julia
